Made api function folder
Moved all API-related functions to their own folder.
