+++ lib/main.dart
@@ -1,82 +1,10 @@
 import 'package:ambience/handlers/file_handler.dart';
 import 'package:ambience/handlers/wallpaper_handler.dart';
-import 'package:ambience/storage/storage.dart';
 import 'package:flutter/material.dart';
-//install http.dart via 'flutter pub add http'
-import 'package:http/http.dart' as http;
-import 'dart:convert';
+import 'package:ambience/api/weather.dart';
 
 void main() {
   runApp(const MyApp());
-}
-
-geolocate(var apiKey, String? input) async {
-  Map coderMap;
-  var cityName = input;
-  //use to limit number of results in geocoder api call
-  //1 means it only takes the top result
-  var limit = 1;
-
-  //geocoder url format
-  //http://api.openweathermap.org/geo/1.0/direct?q={city name},{state code},{country code}&limit={limit}&appid={API key}
-  //country code = ISO 3166 codes, state code is for US only, limit is # of results returned
-
-  var geocodeUrl = Uri.parse(
-      'http://api.openweathermap.org/geo/1.0/direct?q=$cityName&limit=$limit&appid=$apiKey');
-  //api call to geocoder
-  var geocoderReply = await http.get(geocodeUrl);
-
-  //1 size list created by json.decode with map inside if valid
-  var retList = (jsonDecode(geocoderReply.body));
-  //list is empty if given city name could not be resolved
-  if (retList.isEmpty) {
-    return ['failed'];
-  } else {
-    //extract map from list
-    coderMap = retList[0];
-  }
-  return [coderMap['lat'], coderMap['lon']];
-}
-
-getWeather(String? input) async {
-  //to be replaced with a way of hiding api key
-  const apiKey = '91c86752769af03ca919b23664114cda';
-  //cords = 2 size array with lat and lon, index 0 and 1 respectively
-  var cords = await geolocate(apiKey, input);
-  //if name couldn't be geolocated
-  if (cords[0] == 'failed') {
-    debugPrint("Geolocation failed");
-    return false;
-  }
-  //api.openweathermap.org/data/2.5/forecast?lat={lat}&lon={lon}&appid={API key}
-  var weatherUrl = Uri.parse(
-      'https://api.openweathermap.org/data/2.5/forecast?lat=${cords[0]}&lon=${cords[1]}&appid=$apiKey');
-
-  //API call to openweather, sends back a json.
-  var weatherReply = await http.get(weatherUrl);
-  var storage = Storage();
-  //.body attribute contains json, write to given file
-  await storage.writeAppDocFile(weatherReply.body, 'weatherData.txt');
-  //succeeded
-
-  return true;
-}
-
-void weather(String? input) async {
-  //not fully functional
-  //get weather data
-  if (await (getWeather(input)) == false) {
-    debugPrint("Failed to get Weatherdata!");
-  }
-  //else successfuly gotten weather data
-  var storage = Storage();
-  //Reads from give file in AppDoc/Ambience dir and reteurns a decoded JSON as a Dart object
-  var weatherData = await storage.readAppDocJson('weatherData.txt');
-  //sample of parsing weather data
-  //note it is a mix of lists and maps. Hence key and index accesing.
-  debugPrint(
-      "Found weather for ${weatherData['city']['name']} in ${weatherData['city']['country']}."
-      " The weather: ${(weatherData['list'][0])['weather'][0]['description']}.");
 }
 
 class MyApp extends StatelessWidget {
